requirements.txt
# ============================================================================
# SPORTSENSE AI - PYTHON REQUIREMENTS (FULL)
# ============================================================================
# Project: Sportsense AI - Flutter RAG application for sports analytics
# Python: 3.11+ recommended
# ============================================================================
# Installation:
#   pip install -r requirements.txt
#   playwright install
# ============================================================================

# ============================================================================
# CORE WEB FRAMEWORK (API Servers)
# ============================================================================
flask>=3.0.0
flask-cors>=4.0.0

# ============================================================================
# HTTP AND NETWORKING
# ============================================================================
requests>=2.31.0
urllib3>=2.0.0

# ============================================================================
# LLM - LOCAL LANGUAGE MODEL (Qwen2.5-1.5B-Instruct GGUF)
# ============================================================================
llama-cpp-python>=0.2.50

# ============================================================================
# LANGCHAIN - AI ORCHESTRATION & WEB SEARCH (optional)
# ============================================================================
langchain>=0.3.0
langchain-community>=0.3.0
ddgs>=9.0.0

# ============================================================================
# EMBEDDINGS (granite-embedding-278m-multilingual)
# ============================================================================
sentence-transformers>=2.2.0
numpy>=1.24.0

# ============================================================================
# HUGGINGFACE MODEL HUB
# ============================================================================
huggingface_hub>=0.20.0

# ============================================================================
# OPENAI COMPATIBLE CLIENT (for HuggingFace Router API)
# ============================================================================
openai>=1.0.0

# ============================================================================
# WEB BROWSER AUTOMATION (UEFA Rankings Parser)
# ============================================================================
playwright>=1.40.0

# ============================================================================
# HTML PARSING
# ============================================================================
beautifulsoup4>=4.12.0
bs4>=0.0.1
lxml>=4.9.0

# ============================================================================
# UTILITIES
# ============================================================================

# Console output with colors
colorama>=0.4.6

# CLI interface
click>=8.1.0

# Templating engine
jinja2>=3.1.0
markupsafe>=2.1.0

# Async support
anyio>=4.0.0
idna>=3.4
sniffio>=1.3.0

# SSL certificates and encodings
certifi>=2023.0.0
charset_normalizer>=3.3.0

# Path operations
pathlib>=1.0.1

# ============================================================================
# OPTIONAL - DEVELOPMENT AND DEBUGGING
# ============================================================================

# Environment variables from .env file
python-dotenv>=1.0.0

# Progress bars for model downloads
tqdm>=4.66.0

# Data processing
pandas>=2.0.0

# ============================================================================
# PLATFORM-SPECIFIC DEPENDENCIES
# ============================================================================

# Windows
pywin32>=306; sys_platform == 'win32'

# Linux
python-xlib>=0.33; sys_platform == 'linux'

# macOS
pyobjc-core>=10.0; sys_platform == 'darwin'
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'
